// src/app/models/artigo.interface.ts
export interface ArtigosInterface {
    atualizadoPor: string;
    autor: string;
    comentario: string;
    sugestaoComentario: string;
    data: any;
    dataHoraAtualizacao: string | number
    idArtigo: string;
    imageLink: string;
    subTitulo: string;
    isFavorite?: boolean;
    texto: string;
    titulo: string;
    urlOriginal: string;
  }